Store creature base attack power as uint16.

+++ src/server/game/Entities/Creature/Creature.h
@@ -503,8 +503,8 @@ class TC_GAME_API Creature : public Unit, public GridObject<Creature>, public Ma
         uint32 m_assistanceTimer;
 
         // set in Creature::UpdateLevelDependantStats, called in UpdateAttackPowerAndDamage
-        uint32 m_BaseAttackPower;
-        uint32 m_BaseRangedAttackPower;
+        uint16 m_BaseAttackPower;
+        uint16 m_BaseRangedAttackPower;
 };
 
 class TC_GAME_API AssistDelayEvent : public BasicEvent
